+++ src/packages/excalidraw/entry.js
@@ -1,0 +1,6 @@
+import Excalidraw from "./index";
+
+import "../../../public/fonts.css";
+
+export default Excalidraw;
+export * from "./index";